xx_search/requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.1
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
idna==2.9
isort==4.3.21
itchat==1.3.10
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.4
pypng==0.0.20
PyQRCode==1.2.1
pytz==2019.3
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
wrapt==1.11.2
